feynman-platform-whisper-api/requirements.txt
flask==2.3.3
openai-whisper
torch
torchaudio
opencc-python-reimplemented
